refactor: 💡 Door
Refactored and norminetted
