+++ include/bonus_func.h
@@ -6,7 +6,7 @@
 /*   By: hebatist <hebatist@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/14 14:42:16 by hebatist          #+#    #+#             */
-/*   Updated: 2025/08/31 15:40:45 by hebatist         ###   ########.fr       */
+/*   Updated: 2025/08/31 21:03:11 by hebatist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,5 +29,12 @@ void	invert_drawing(t_mlx *st_mlx, int x);
 t_spr	*build_orb(t_mlx *st_mlx, char c);
 void	set_orb(t_mlx *st_mlx);
 void	put_orb_pixel(t_mlx *st_mlx, t_spr *st_spr, int y, int line);
+void	set_door_coord(t_mlx *st_mlx, int *curr_map_x, int *curr_map_y);
+void	start_door_anim(t_mlx *st_mlx);
+int		set_ceiling_color(t_mlx *st_mlx);
+int		set_floor_color(t_mlx *st_mlx);
+void	ft_mlx_pixel_put(t_img *st_img, int x, int y, int color);
+void	put_door_texture_pixel(t_mlx *st_mlx, int screen_column, int y);
+void	put_wall_texture_pixel(t_mlx *st_mlx, int screen_column, int y);
 
 #endif
