+++ src/bonus/gameplay_effects_bonus.c
@@ -6,7 +6,7 @@
 /*   By: hebatist <hebatist@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/22 03:50:44 by hebatist          #+#    #+#             */
-/*   Updated: 2025/08/31 15:59:57 by hebatist         ###   ########.fr       */
+/*   Updated: 2025/08/31 20:49:21 by hebatist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,5 +72,31 @@ void	start_curtain_effect(t_mlx *st_mlx)
 	{
 		st_mlx->curtain_dir = 0;
 		st_mlx->is_curtain = 0;
+	}
+}
+
+void	start_door_anim(t_mlx *st_mlx)
+{
+	if (st_mlx->door_dir == 0)
+	{
+		if (st_mlx->door_offset < 100)
+			st_mlx->door_offset += 1;
+		else
+		{
+			st_mlx->is_door_anim = 0;
+			st_mlx->door_dir = 1;
+			st_mlx->is_door_open = 1;
+		}
+	}
+	else
+	{
+		if (st_mlx->door_offset > 0)
+			st_mlx->door_offset -= 1;
+		else
+		{
+			st_mlx->is_door_anim = 0;
+			st_mlx->door_dir = 0;
+			st_mlx->is_door_open = 0;
+		}
 	}
 }
