+++ src/bonus/gameplay_utils1_bonus.c
@@ -1,53 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   gameplay_utils_bonus.c                             :+:      :+:    :+:   */
+/*   gameplay_utils1_bonus.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: dbatista <dbatista@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/05 22:26:51 by dbatista          #+#    #+#             */
-/*   Updated: 2025/08/30 22:55:12 by hebatist         ###   ########.fr       */
+/*   Updated: 2025/08/31 21:57:24 by hebatist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/cub3d_bonus.h"
-
-int	is_not_valid_move(double new_x, double new_y, t_mlx *st_mlx)
-{
-	t_file	*st_file;
-	char	**map;
-	int		map_x;
-	int		map_y;
-
-	st_file = st_mlx->st_file;
-	map = st_file->map;
-	map_x = (int)new_x;
-	map_y = (int)new_y;
-	if (new_x < 0 || new_y < 0)
-		return (1);
-	if (map_y < 0 || map_y >= st_file->height)
-		return (1);
-	if (map_x < 0 || map_x >= (int)ft_strlen(map[map_y]))
-		return (1);
-	if (map[(int)(new_y + RADIUS)][(int)(new_x + RADIUS)] == '1'
-	|| map[(int)(new_y - RADIUS)][(int)(new_x - RADIUS)] == '1'
-	|| map[(int)(new_y + RADIUS)][(int)(new_x - RADIUS)] == '1'
-	|| map[(int)(new_y - RADIUS)][(int)(new_x + RADIUS)] == '1')
-		return (1);
-	if (map[map_y][map_x] == '1' || map[map_y][map_x] == ' ')
-		return (1);
-	return (0);
-}
-
-void	move_player(t_mlx *st_mlx, double move_x, double move_y)
-{
-	if (!is_not_valid_move(st_mlx->st_coord->p_posx + move_x,
-			st_mlx->st_coord->p_posy, st_mlx))
-		st_mlx->st_coord->p_posx += move_x;
-	if (!is_not_valid_move(st_mlx->st_coord->p_posx,
-			st_mlx->st_coord->p_posy + move_y, st_mlx))
-		st_mlx->st_coord->p_posy += move_y;
-}
 
 void	try_open_door(t_mlx *st_mlx)
 {
@@ -57,10 +20,7 @@ void	try_open_door(t_mlx *st_mlx)
 	check_x = (int)(st_mlx->st_coord->p_posx + st_mlx->st_coord->dir_vec_x);
 	check_y = (int)(st_mlx->st_coord->p_posy + st_mlx->st_coord->dir_vec_y);
 	if (st_mlx->st_file->map[check_y][check_x] == 'D')
-	{
 		st_mlx->is_door_anim = 1;
-		printf("check_x=%d check_y=%d cell=%c open=%d\n", check_x, check_y, st_mlx->st_file->map[check_y][check_x], st_mlx->is_door_open);
-	}
 }
 
 int	key_press(int key, t_mlx *st_mlx)
@@ -85,10 +45,7 @@ int	key_press(int key, t_mlx *st_mlx)
 		else if (key == KEY_RIGHT)
 			st_mlx->key_right = 1;
 		else if (key == KEY_SPACE)
-		{
-			printf("teclas apertada: %d\n", key);
 			try_open_door(st_mlx);
-		}
 	}
 	return (0);
 }
